YYC-108: extract tui events + keymap submodules
Continues YYC-108. Pulls two more layers out of `tui/mod.rs`:

- src/tui/events.rs (NEW, 218 lines): stream-event handling.
  Owns handle_stream_event (the per-event match), submit_prompt
  + refresh_sessions (the agent-launch / session-refresh helpers
  the Done branch and Enter handler share),
  render_wake_for_stream_batch (per-frame draw budget),
  stream_event_forces_redraw, RenderWake, STREAM_FRAME_BUDGET.
- src/tui/keymap.rs (NEW, 236 lines): slash-command surface.
  Owns SlashCommand + the SLASH_COMMANDS const table,
  filter_commands, current_palette, complete_slash,
  format_model_list / format_provider_list (kept around for
  --help-style printers and tests),
  build_provider_picker_entries.

src/tui/mod.rs: 1940 → 1526 lines. Module declarations updated
(`mod events; mod keymap;`); call sites switched to `events::`
and `keymap::` qualifiers. Unused `Duration` import dropped from
mod.rs and added to the test sub-module where it's still used.
src/tui/rendering.rs's `super::SlashCommand` import re-pointed to
`super::keymap::SlashCommand`.

Combined with the rendering + init extraction in the previous
commit on this branch, the running totals are:

  Original  src/tui/mod.rs  2570 lines
  After     src/tui/mod.rs  1526 lines  (-1044)

Public API + behavior unchanged. All 355 lib tests pass; clippy
clean.

The original YYC-108 acceptance ("drop below ~1000 lines") is
not met yet — `run_tui` itself is still ~900 lines of input
dispatch + main-loop wiring. That extraction (likely
`tui/dispatch.rs` for the slash-command match arms +
`tui/input.rs` for the key-event branches) is a follow-up; this
commit lands the natural decomposition seams without rewriting
the loop body.
