+++ src/tui/mod.rs
@@ -1,5 +1,5 @@
 use std::sync::Arc;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use anyhow::Result;
 use ratatui::{
@@ -18,8 +18,10 @@ use crate::provider::StreamEvent;
 
 pub mod chat_message;
 pub mod chat_render;
+mod events;
 mod init;
 pub mod keybinds;
+mod keymap;
 pub mod markdown;
 pub mod miller_columns;
 pub mod model_picker;
@@ -34,35 +36,6 @@ pub mod widgets;
 use state::{AppState, ChatMessage, ChatRole};
 use theme::{Theme, body};
 use views::{View, render_view};
-
-const STREAM_FRAME_BUDGET: Duration = Duration::from_millis(16);
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum RenderWake {
-    Now,
-    Wait(Duration),
-}
-
-fn render_wake_for_stream_batch(
-    last_draw: Instant,
-    now: Instant,
-    is_terminal_event: bool,
-) -> RenderWake {
-    if is_terminal_event {
-        return RenderWake::Now;
-    }
-
-    let elapsed = now.saturating_duration_since(last_draw);
-    if elapsed >= STREAM_FRAME_BUDGET {
-        RenderWake::Now
-    } else {
-        RenderWake::Wait(STREAM_FRAME_BUDGET - elapsed)
-    }
-}
-
-fn stream_event_forces_redraw(ev: &StreamEvent) -> bool {
-    matches!(ev, StreamEvent::Done(_) | StreamEvent::Error(_))
-}
 
 /// What session, if any, the TUI should load on startup.
 #[derive(Debug, Clone)]
@@ -84,397 +57,8 @@ enum KeyEv {
     Error(String),
 }
 
-#[derive(Debug, Clone)]
-struct SlashCommand {
-    name: &'static str,
-    description: &'static str,
-    /// True when the command can run mid-turn without corrupting agent state
-    /// (YYC-62). Pure UI ops are safe; anything that mutates conversation
-    /// history or reaches into the agent is not. Default false (conservative).
-    mid_turn_safe: bool,
-}
-
-const SLASH_COMMANDS: &[SlashCommand] = &[
-    SlashCommand {
-        name: "exit",
-        description: "Quit Vulcan",
-        // Always exits cleanly; no state to corrupt.
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "quit",
-        description: "Quit Vulcan",
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "help",
-        description: "Show available commands",
-        // Pure UI: pushes a system message.
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "clear",
-        description: "Clear message history",
-        // Destructive: would nuke the in-flight User+Agent pair the agent
-        // loop is streaming into. Defer until idle.
-        mid_turn_safe: false,
-    },
-    SlashCommand {
-        name: "view",
-        description: "Cycle to next view (or 1-5)",
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "reasoning",
-        description: "Toggle reasoning trace",
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "search",
-        description: "Search past sessions: /search <query>",
-        // Holds agent.lock().await — would deadlock against the in-flight
-        // run_prompt_stream task. Defer until idle.
-        mid_turn_safe: false,
-    },
-    SlashCommand {
-        name: "model",
-        description: "List or switch models: /model [id]",
-        // Rebuilds the provider for future turns and may fetch the catalog.
-        // Defer until idle so the in-flight provider stream is untouched.
-        mid_turn_safe: false,
-    },
-    SlashCommand {
-        name: "provider",
-        description: "List or switch named providers: /provider [name|default]",
-        // Rebuilds the provider against a different profile; same idle
-        // requirement as /model.
-        mid_turn_safe: false,
-    },
-    SlashCommand {
-        name: "diff-style",
-        description: "Set diff render: /diff-style <unified|side-by-side|inline>",
-        mid_turn_safe: true,
-    },
-    SlashCommand {
-        name: "resume",
-        description: "Open session picker to switch to another session",
-        mid_turn_safe: false,
-    },
-];
-
 fn short_id(id: &str) -> String {
     id.chars().take(8).collect()
-}
-
-#[allow(dead_code)] // retained for tests and potential `--help`-style printers.
-fn format_model_list(active_model: &str, models: &[crate::provider::catalog::ModelInfo]) -> String {
-    let mut out = format!("Models from active provider ({} total):", models.len());
-    for model in models.iter().take(30) {
-        let marker = if model.id == active_model { "*" } else { " " };
-        let context = if model.context_length > 0 {
-            crate::tui::state::format_thousands(model.context_length as u32)
-        } else {
-            "unknown".into()
-        };
-        let mut flags = Vec::new();
-        if model.features.tools {
-            flags.push("tools");
-        }
-        if model.features.reasoning {
-            flags.push("reasoning");
-        }
-        if model.features.vision {
-            flags.push("vision");
-        }
-        if model.features.json_mode {
-            flags.push("json");
-        }
-        let flags = if flags.is_empty() {
-            String::new()
-        } else {
-            format!(" · {}", flags.join(","))
-        };
-        out.push_str(&format!("\n  {marker} {} · ctx {context}{flags}", model.id));
-    }
-    if models.len() > 30 {
-        out.push_str(&format!("\n  ... {} more", models.len() - 30));
-    }
-    out.push_str("\n\nUse /model <id> to switch.");
-    out
-}
-
-fn build_provider_picker_entries(config: &Config) -> Vec<crate::tui::state::ProviderPickerEntry> {
-    use crate::tui::state::ProviderPickerEntry;
-    let mut out = Vec::with_capacity(config.providers.len() + 1);
-    out.push(ProviderPickerEntry {
-        name: None,
-        model: config.provider.model.clone(),
-        base_url: config.provider.base_url.clone(),
-    });
-    let mut names: Vec<&String> = config.providers.keys().collect();
-    names.sort();
-    for name in names {
-        let cfg = &config.providers[name];
-        out.push(ProviderPickerEntry {
-            name: Some(name.clone()),
-            model: cfg.model.clone(),
-            base_url: cfg.base_url.clone(),
-        });
-    }
-    out
-}
-
-#[allow(dead_code)] // retained for tests and potential `--help`-style printers.
-fn format_provider_list(config: &Config, active: Option<&str>) -> String {
-    let mut out = String::from("Provider profiles:");
-    let default_marker = if active.is_none() { "*" } else { " " };
-    out.push_str(&format!(
-        "\n  {default_marker} default · {} · {}",
-        config.provider.base_url, config.provider.model,
-    ));
-
-    let mut names: Vec<&String> = config.providers.keys().collect();
-    names.sort();
-    for name in names {
-        let cfg = &config.providers[name];
-        let marker = if active == Some(name.as_str()) {
-            "*"
-        } else {
-            " "
-        };
-        out.push_str(&format!(
-            "\n  {marker} {name} · {} · {}",
-            cfg.base_url, cfg.model,
-        ));
-    }
-    if config.providers.is_empty() {
-        out.push_str("\n  (no named [providers.<name>] profiles configured)");
-    }
-    out.push_str("\n\nUse /provider <name> to switch, /provider default to revert.");
-    out
-}
-
-fn filter_commands(prefix: &str) -> Vec<&'static SlashCommand> {
-    if prefix.is_empty() {
-        return SLASH_COMMANDS.iter().collect();
-    }
-    let lower = prefix.to_lowercase();
-    SLASH_COMMANDS
-        .iter()
-        .filter(|c| c.name.starts_with(&lower))
-        .collect()
-}
-
-/// Same matching logic as the palette renderer in the main loop — exposed
-/// as a helper so the key handler can decide what to highlight or commit
-/// without duplicating prefix logic (YYC-70).
-fn current_palette(input: &str) -> Vec<&'static SlashCommand> {
-    if input == "/" {
-        SLASH_COMMANDS.iter().collect()
-    } else if input.starts_with('/') && input.len() > 1 {
-        filter_commands(&input[1..])
-    } else {
-        Vec::new()
-    }
-}
-
-fn complete_slash(prefix: &str) -> Option<String> {
-    let matches = filter_commands(prefix);
-    if matches.is_empty() {
-        return None;
-    }
-    if matches.len() == 1 {
-        return Some(matches[0].name.to_string());
-    }
-    let first = matches[0].name.as_bytes();
-    let mut common = first.len();
-    for m in &matches[1..] {
-        let bytes = m.name.as_bytes();
-        common = common.min(bytes.len());
-        for (i, &b) in first.iter().enumerate().take(common) {
-            if b != bytes[i] {
-                common = i;
-                break;
-            }
-        }
-    }
-    if common > prefix.len() {
-        Some(matches[0].name[..common].to_string())
-    } else {
-        None
-    }
-}
-
-/// Spawn a fresh agent turn for `msg`. Updates chat state (User + empty
-/// Agent message), flips thinking on, re-engages auto-follow, then spawns
-/// `run_prompt_stream` against the agent. Used by the Enter handler for
-/// new submissions and by the Done handler when draining the queue
-/// (YYC-61).
-fn submit_prompt(
-    app: &mut AppState,
-    agent: &Arc<Mutex<Agent>>,
-    stream_tx: &mpsc::Sender<StreamEvent>,
-    msg: String,
-) {
-    app.messages.push(ChatMessage {
-        role: ChatRole::User,
-        content: msg.clone(),
-        ..Default::default()
-    });
-    app.messages.push(ChatMessage {
-        role: ChatRole::Agent,
-        content: String::new(),
-        ..Default::default()
-    });
-    app.thinking = true;
-    app.at_bottom = true;
-    app.note_prompt_submitted(&msg);
-
-    // YYC-105: hold a cancel token outside the agent mutex so Ctrl+C
-    // can fire it without waiting for the prompt task to release the
-    // lock. The agent mirrors the same token internally for tools /
-    // hooks that still consult `self.turn_cancel`.
-    let cancel = tokio_util::sync::CancellationToken::new();
-    app.current_turn_cancel = Some(cancel.clone());
-    let tx = stream_tx.clone();
-    let a = agent.clone();
-    tokio::spawn(async move {
-        let mut a = a.lock().await;
-        let _ = a.run_prompt_stream_with_cancel(&msg, tx, cancel).await;
-    });
-}
-
-async fn refresh_sessions(agent: &Arc<Mutex<Agent>>, app: &mut AppState) {
-    let (summaries, active_session_id) = {
-        let a = agent.lock().await;
-        (
-            a.memory().list_sessions(12).unwrap_or_default(),
-            a.session_id().to_string(),
-        )
-    };
-    app.hydrate_sessions(&summaries, &active_session_id);
-}
-
-async fn handle_stream_event(
-    app: &mut AppState,
-    agent: &Arc<Mutex<Agent>>,
-    stream_tx: &mpsc::Sender<StreamEvent>,
-    ev: StreamEvent,
-) {
-    match ev {
-        StreamEvent::Text(chunk) => {
-            if let Some(last) = app.messages.last_mut()
-                && matches!(last.role, ChatRole::Agent)
-            {
-                // Append to both the segment timeline (the YYC-71
-                // ordered renderer) and the legacy `content` field
-                // (kept so other code that peeks at .content keeps working).
-                last.append_text(&chunk);
-                // Strip leading whitespace so models that emit `\n\n`
-                // preambles don't render gaps before the visible body
-                // when the renderer falls back to `content`.
-                if last.content.is_empty() {
-                    let trimmed = chunk.trim_start_matches(|c: char| {
-                        c == '\n' || c == '\r' || c == ' ' || c == '\t'
-                    });
-                    last.content.push_str(trimmed);
-                } else {
-                    last.content.push_str(&chunk);
-                }
-            }
-        }
-        StreamEvent::Reasoning(chunk) => {
-            // Per-token reasoning trace from thinking-mode models. Push to
-            // the segment timeline so it interleaves with tool calls in
-            // render order; also append to the legacy `reasoning` field so
-            // latest_reasoning() etc. continue to work.
-            if let Some(last) = app.messages.last_mut()
-                && matches!(last.role, ChatRole::Agent)
-            {
-                last.append_reasoning(&chunk);
-                last.reasoning.push_str(&chunk);
-            }
-            app.note_reasoning();
-        }
-        StreamEvent::Done(resp) => {
-            app.thinking = false;
-            app.current_turn_cancel = None;
-            if let Some(usage) = resp.usage {
-                // YYC-60: track lifetime totals for cost (YYC-67) and the
-                // latest prompt size for the in-status capacity bar.
-                app.prompt_tokens_total = app
-                    .prompt_tokens_total
-                    .saturating_add(usage.prompt_tokens as u32);
-                app.completion_tokens_total = app
-                    .completion_tokens_total
-                    .saturating_add(usage.completion_tokens as u32);
-                app.prompt_tokens_last = usage.prompt_tokens as u32;
-            }
-            app.note_done();
-            refresh_sessions(agent, app).await;
-            // YYC-61: drain one queued prompt per turn end. Subsequent queued
-            // prompts ride the next Done event in the same way.
-            if let Some(next) = app.queue.pop_front() {
-                submit_prompt(app, agent, stream_tx, next);
-            }
-        }
-        StreamEvent::Error(e) => {
-            if let Some(last) = app.messages.last_mut()
-                && last.content.is_empty()
-            {
-                last.set_content(format!("⚠ Error: {e}"));
-            }
-            app.thinking = false;
-            app.current_turn_cancel = None;
-            // YYC-67: record provider-level error for telemetry.
-            app.provider_errors_total = app.provider_errors_total.saturating_add(1);
-            app.note_error(&e);
-        }
-        StreamEvent::ToolCallStart {
-            name, args_summary, ..
-        } => {
-            // YYC-71: push the tool-call segment into the timeline
-            // (interleaved with reasoning/text). YYC-74: carry the args
-            // summary so the card has structured context.
-            if let Some(last) = app.messages.last_mut()
-                && matches!(last.role, ChatRole::Agent)
-            {
-                last.push_tool_start_with(name.clone(), args_summary);
-            }
-            app.note_tool_start(&name);
-        }
-        StreamEvent::ToolCallEnd {
-            name,
-            ok,
-            output_preview,
-            result_meta,
-            elided_lines,
-            elapsed_ms,
-            ..
-        } => {
-            if let Some(last) = app.messages.last_mut()
-                && matches!(last.role, ChatRole::Agent)
-            {
-                // YYC-74: stamp preview + meta + timing onto the matching
-                // segment for the card. YYC-78: stash elided count for the
-                // collapse footer.
-                last.finish_tool_with(
-                    &name,
-                    ok,
-                    output_preview,
-                    result_meta,
-                    elided_lines,
-                    Some(elapsed_ms),
-                );
-            }
-            // YYC-67: tool call telemetry.
-            app.tool_calls_total = app.tool_calls_total.saturating_add(1);
-            if !ok {
-                app.tool_errors_total = app.tool_errors_total.saturating_add(1);
-            }
-            app.note_tool_end(&name, ok);
-        }
-    }
 }
 
 pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
@@ -581,7 +165,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
         app.token_max = a.max_context() as u32;
         app.provider_label = a.active_profile().map(str::to_string);
     }
-    refresh_sessions(&agent, &mut app).await;
+    events::refresh_sessions(&agent, &mut app).await;
 
     // YYC-86: if the user invoked --resume, activate the session picker.
     app.show_session_picker = matches!(resume, ResumeTarget::Pick);
@@ -638,9 +222,9 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
 
     while !exit {
         let palette = if app.input.starts_with('/') && app.input.len() > 1 {
-            Some(filter_commands(&app.input[1..]))
+            Some(keymap::filter_commands(&app.input[1..]))
         } else if app.input == "/" {
-            Some(SLASH_COMMANDS.iter().collect())
+            Some(keymap::SLASH_COMMANDS.iter().collect())
         } else {
             None
         };
@@ -1096,7 +680,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                                         }
                                                     }
                                                 }
-                                                refresh_sessions(&agent, &mut app).await;
+                                                events::refresh_sessions(&agent, &mut app).await;
                                             }
                                         }
                                         KeyCode::Esc => {
@@ -1267,7 +851,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                 if app.keybinds.toggle_sessions.matches(&key)
                                     && !app.input.starts_with('/')
                                 {
-                                    refresh_sessions(&agent, &mut app).await;
+                                    events::refresh_sessions(&agent, &mut app).await;
                                     app.show_session_picker = true;
                                     app.session_picker_selection = 0;
                                     continue;
@@ -1302,7 +886,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                         // slash menu when it's open.
                                         KeyCode::Char('j') | KeyCode::Char('k') => {
                                             if app.input.starts_with('/') {
-                                                let candidates = current_palette(&app.input);
+                                                let candidates = keymap::current_palette(&app.input);
                                                 if !candidates.is_empty() {
                                                     let len = candidates.len();
                                                     if matches!(key.code, KeyCode::Char('j')) {
@@ -1356,7 +940,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                         // YYC-70: when the slash menu is open with at least one
                                         // match, Enter commits the highlighted command.
                                         if app.input.starts_with('/') {
-                                            let candidates = current_palette(&app.input);
+                                            let candidates = keymap::current_palette(&app.input);
                                             if !candidates.is_empty() {
                                                 let idx = app.slash_menu_selection.min(candidates.len() - 1);
                                                 app.input = format!("/{}", candidates[idx].name);
@@ -1374,7 +958,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                         let mid_turn_safe = if is_slash {
                                             let body = &app.input[1..];
                                             let head = body.split_whitespace().next().unwrap_or("");
-                                            SLASH_COMMANDS
+                                            keymap::SLASH_COMMANDS
                                                 .iter()
                                                 .find(|c| c.name == head)
                                                 .map(|c| c.mid_turn_safe)
@@ -1424,7 +1008,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                                     "exit" | "quit" => { exit = true; continue; }
                                                     "help" => {
                                                         let mut help = String::from("Commands:");
-                                                        for cmd in SLASH_COMMANDS {
+                                                        for cmd in keymap::SLASH_COMMANDS {
                                                             help.push_str(&format!("\n  /{:<10}  {}", cmd.name, cmd.description));
                                                         }
                                                         help.push_str("\n\nKeys:\n  Ctrl+1..5  switch view (1=stack 2=split 3=tiled 4=tree 5=floor)\n  Ctrl+R     toggle reasoning trace\n  Tab        complete slash command");
@@ -1495,7 +1079,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                                         continue;
                                                     }
                                                     "resume" => {
-                                                        refresh_sessions(&agent, &mut app).await;
+                                                        events::refresh_sessions(&agent, &mut app).await;
                                                         app.show_session_picker = true;
                                                         app.session_picker_selection = 0;
                                                         continue;
@@ -1551,7 +1135,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                                                 a.active_profile().map(str::to_string)
                                                             };
                                                             app.provider_picker_items =
-                                                                build_provider_picker_entries(config);
+                                                                keymap::build_provider_picker_entries(config);
                                                             let active_idx = app
                                                                 .provider_picker_items
                                                                 .iter()
@@ -1680,7 +1264,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                                 }
                                             }
 
-                                            submit_prompt(&mut app, &agent, &stream_tx, msg);
+                                            events::submit_prompt(&mut app, &agent, &stream_tx, msg);
                                         }
                                     }
                                     KeyCode::Char(c) => {
@@ -1698,14 +1282,14 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                     KeyCode::Tab => {
                                         pending_quit = false;
                                         if let Some(rest) = app.input.strip_prefix('/')
-                                            && let Some(c) = complete_slash(rest) {
+                                            && let Some(c) = keymap::complete_slash(rest) {
                                                 app.input = format!("/{c}");
                                             }
                                     }
                                     KeyCode::Up => {
                                         // YYC-70: arrows navigate the slash menu when open.
                                         if app.input.starts_with('/') {
-                                            let candidates = current_palette(&app.input);
+                                            let candidates = keymap::current_palette(&app.input);
                                             if !candidates.is_empty() {
                                                 app.slash_menu_selection =
                                                     app.slash_menu_selection.saturating_sub(1);
@@ -1717,7 +1301,7 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
                                     }
                                     KeyCode::Down => {
                                         if app.input.starts_with('/') {
-                                            let candidates = current_palette(&app.input);
+                                            let candidates = keymap::current_palette(&app.input);
                                             if !candidates.is_empty() {
                                                 let len = candidates.len();
                                                 app.slash_menu_selection =
@@ -1763,15 +1347,15 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
             ev = stream_rx.recv() => {
                 match ev {
                     Some(ev) => {
-                        let mut force_redraw = stream_event_forces_redraw(&ev);
-                        handle_stream_event(&mut app, &agent, &stream_tx, ev).await;
+                        let mut force_redraw = events::stream_event_forces_redraw(&ev);
+                        events::handle_stream_event(&mut app, &agent, &stream_tx, ev).await;
                         while let Ok(ev) = stream_rx.try_recv() {
-                            force_redraw |= stream_event_forces_redraw(&ev);
-                            handle_stream_event(&mut app, &agent, &stream_tx, ev).await;
+                            force_redraw |= events::stream_event_forces_redraw(&ev);
+                            events::handle_stream_event(&mut app, &agent, &stream_tx, ev).await;
                         }
                         if !force_redraw
-                            && let RenderWake::Wait(delay) =
-                                render_wake_for_stream_batch(last_draw, Instant::now(), false)
+                            && let events::RenderWake::Wait(delay) =
+                                events::render_wake_for_stream_batch(last_draw, Instant::now(), false)
                             {
                                 tokio::time::sleep(delay).await;
                             }
@@ -1799,14 +1383,15 @@ pub async fn run_tui(config: &Config, resume: ResumeTarget) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::time::Duration;
 
     #[test]
     fn stream_batching_caps_stream_redraws_to_frame_budget() {
         let start = Instant::now();
 
         assert_eq!(
-            render_wake_for_stream_batch(start, start + Duration::from_millis(1), false),
-            RenderWake::Wait(Duration::from_millis(15))
+            events::render_wake_for_stream_batch(start, start + Duration::from_millis(1), false),
+            events::RenderWake::Wait(Duration::from_millis(15))
         );
     }
 
@@ -1815,20 +1400,20 @@ mod tests {
         let start = Instant::now();
 
         assert_eq!(
-            render_wake_for_stream_batch(start, start + Duration::from_millis(1), true),
-            RenderWake::Now
+            events::render_wake_for_stream_batch(start, start + Duration::from_millis(1), true),
+            events::RenderWake::Now
         );
     }
 
     #[test]
     fn model_command_is_available_and_deferred_mid_turn() {
-        let command = SLASH_COMMANDS
+        let command = keymap::SLASH_COMMANDS
             .iter()
             .find(|cmd| cmd.name == "model")
             .expect("model slash command");
 
         assert!(!command.mid_turn_safe);
-        assert_eq!(filter_commands("mod")[0].name, "model");
+        assert_eq!(keymap::filter_commands("mod")[0].name, "model");
     }
 
     #[test]
@@ -1851,7 +1436,7 @@ mod tests {
         config.provider.model = "deepseek/v4".into();
         config.providers = providers;
 
-        let entries = build_provider_picker_entries(&config);
+        let entries = keymap::build_provider_picker_entries(&config);
         assert_eq!(entries.len(), 3);
         assert!(entries[0].name.is_none());
         assert_eq!(entries[0].model, "deepseek/v4");
@@ -1861,13 +1446,13 @@ mod tests {
 
     #[test]
     fn provider_command_is_available_and_deferred_mid_turn() {
-        let command = SLASH_COMMANDS
+        let command = keymap::SLASH_COMMANDS
             .iter()
             .find(|cmd| cmd.name == "provider")
             .expect("provider slash command");
 
         assert!(!command.mid_turn_safe);
-        assert_eq!(filter_commands("prov")[0].name, "provider");
+        assert_eq!(keymap::filter_commands("prov")[0].name, "provider");
     }
 
     #[test]
@@ -1886,11 +1471,11 @@ mod tests {
         config.provider.model = "deepseek/v4".into();
         config.providers = providers;
 
-        let active_default = format_provider_list(&config, None);
+        let active_default = keymap::format_provider_list(&config, None);
         assert!(active_default.contains("* default · https://openrouter.ai/api/v1 · deepseek/v4"));
         assert!(active_default.contains("  local · http://localhost:11434/v1 · qwen2.5"));
 
-        let active_local = format_provider_list(&config, Some("local"));
+        let active_local = keymap::format_provider_list(&config, Some("local"));
         assert!(active_local.contains("  default · https://openrouter.ai/api/v1"));
         assert!(active_local.contains("* local · http://localhost:11434/v1"));
     }
@@ -1899,7 +1484,7 @@ mod tests {
     fn format_provider_list_handles_no_named_profiles() {
         use crate::config::Config;
         let config = Config::default();
-        let report = format_provider_list(&config, None);
+        let report = keymap::format_provider_list(&config, None);
         assert!(report.contains("* default"));
         assert!(report.contains("(no named [providers.<name>] profiles configured)"));
     }
@@ -1930,7 +1515,7 @@ mod tests {
             },
         ];
 
-        let report = format_model_list("model-a", &models);
+        let report = keymap::format_model_list("model-a", &models);
 
         assert!(report.contains("* model-a · ctx 1,000 · tools,json"));
         assert!(report.contains("  model-b · ctx unknown"));
